[level 1] Title: 정수 제곱근 판별, Time: 0.05 ms, Memory: 72 MB -BaekjoonHub

+++ 프로그래머스/lv1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -2,7 +2,7 @@ class Solution {
     public long solution(long n) {
        long answer = 0;
         double sqrt = Math.sqrt(n);
-        if(sqrt==(long)sqrt){
+        if(sqrt==(int)sqrt){
             answer = (long)Math.pow(sqrt+1,2);
         }else{
             answer = -1;
